Rework type system a bit for array-adapter:
The array adapter interface accepts a type <T>, just like any other
array which wants to know what it is an array of. Previously, there had
been enforcements of SupportedDataPointType as the type of T, which
still isn't a terrible idea, but a) probably won't matter unless users
are using typescript importing types and b) makes testing validation
inconvenient because certain tests validate vanilla number arrays
(number[]) instead of SupportedDataPointType[], as that is a shortcut to
SimpleDataPoint (a supported type).